# openapi.yaml
openapi: 3.1.0
info:
  title: 6MD Management Platform API
  description: |
    Web application API for managing Arma 3 unit operations, signups, attendance, ranks, 
    training progression, and admin workflows for the 6th Maroon Division.
  version: 1.0.0
  contact:
    name: 6th Maroon Division
  license:
    name: Proprietary
    identifier: 6MD
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://orbat.6md.net/api
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /orbats:
    get:
      tags:
        - ORBATs
      summary: List recent ORBATs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: List of ORBATs
    post:
      tags:
        - ORBATs
      summary: Create new ORBAT
      security:
        - bearerAuth:
            - orbat:create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrbatCreateInput"
      responses:
        "201":
          description: ORBAT created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
  /ranks/auto-rankup:
    post:
      tags:
        - Ranks
      summary: Check auto-rankup eligibility
      responses:
        "200":
          description: Auto-rankup eligibility check result
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth.js session token (cookie-based)
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: BotToken
      description: Active bot token stored in the BotToken database table. Use Authorization Bearer; environment tokens and X-BOT-API-TOKEN are not accepted.
  schemas:
...
    OrbatCreateInput:
      type: object
      required:
        - name
        - squads
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        eventDate:
          type: string
          format: date
          nullable: true
        startTime:
          type: string
          format: time
          nullable: true
        endTime:
          type: string
          format: time
          nullable: true
        squads:
          type: array
          items:
            type: object
            required:
              - name
              - slots
            properties:
              name:
                type: string
              orderIndex:
                type: integer
              slots:
                type: array
                items:
                  type: object
                  required:
                    - name
                    - orderIndex
                    - maxSignups
                  properties:
                    name:
                      type: string
                    orderIndex:
                      type: integer
                    maxSignups:
                      type: integer
                    squadRoleId:
                      type: integer
                      nullable: true
        frequencyIds:
          type: array
          items:
            type: integer
        tempFrequencies:
          type: array
          items:
            type: object
            properties:
              frequency:
                type: string
              type:
                type: string
                enum:
                  - SR
                  - LR
              isAdditional:
                type: boolean
              channel:
                type: string
              callsign:
                type: string
        bluforCountry:
          type: string
          nullable: true
        opforCountry:
          type: string
          nullable: true
        indepCountry:
          type: string
          nullable: true
        rulesOfEngagement:
          type: string
          nullable: true
...